Move DeferredProxy<T> to api/deferred-proxy.h
I switched io-context.h to include api/deferred-proxy.h instead of api/util.h, which necessitated some explicit inclusions of api/util.h in some .c++ files.
