+++ src/workerd/api/deferred-proxy-test.c++
@@ -1,4 +1,4 @@
-#include "util.h"
+#include "deferred-proxy.h"
 #include <kj/test.h>
 
 namespace workerd::api {
